Move contents of _Column into _Cubes
